renamed file: VarsRequiredFromSubtree.h -> ColumnStrippingHelpers.h
