+++ src/engine/ExistsJoin.cpp
@@ -16,7 +16,7 @@
 #include "util/ChunkedForLoop.h"
 #include "util/JoinAlgorithms/IndexNestedLoopJoin.h"
 #include "util/JoinAlgorithms/JoinAlgorithms.h"
-#include "util/VarsRequiredFromSubtree.h"
+#include "util/ColumnStrippingHelpers.h"
 #include "util/VectorWithMemoryLimit.h"
 
 // _____________________________________________________________________________
