+++ src/util/ColumnStrippingHelpers.h
